camel-security/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "camel-security"
version = "1.0.0"
description = "CaMeL: Capabilities for Machine Learning - A defense system against prompt injection attacks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "CaMeL Implementation", email = "camel@example.com"}
]
keywords = [
    "llm",
    "security",
    "prompt-injection",
    "capabilities",
    "agentic-systems",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.18.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "pyyaml>=6.0",
]
all = [
    "camel-security[anthropic,dev]",
]

[project.urls]
Homepage = "https://github.com/example/camel-security"
Documentation = "https://github.com/example/camel-security#readme"
Repository = "https://github.com/example/camel-security"

[tool.setuptools.packages.find]
include = ["camel*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
